feat: add entity exist check in edit and delete, use more specialized exceptions in CourtSurfaceServiceImpl

+++ src/main/java/com/example/inqooltennisreservationapi/service/impl/CourtSurfaceServiceImpl.java
@@ -1,5 +1,7 @@
 package com.example.inqooltennisreservationapi.service.impl;
 
+import com.example.inqooltennisreservationapi.exceptions.DatabaseException;
+import com.example.inqooltennisreservationapi.exceptions.EntityNotFoundException;
 import com.example.inqooltennisreservationapi.model.api.CourtSurfaceDTOs;
 import com.example.inqooltennisreservationapi.model.mappers.CourtSurfaceMapper;
 import com.example.inqooltennisreservationapi.repository.CourtSurfaceRepository;
@@ -25,7 +27,7 @@ public class CourtSurfaceServiceImpl implements CourSurfaceService {
     public CourtSurfaceDTOs.CourtSurfaceResponseDTO getSurface(long id) {
         var entity = surfaceRepository.getCourtSurfaceById(id);
         if (entity.isEmpty()) {
-            throw new RuntimeException("No such entity");
+            throw new EntityNotFoundException(String.format("Court surface with id %s not found", id));
         }
         return courtSurfaceMapper.entityToResponseDto(entity.get());
     }
@@ -37,7 +39,7 @@ public class CourtSurfaceServiceImpl implements CourSurfaceService {
 
         var result = surfaceRepository.createCourtSurface(entityToSave);
         if (result.isEmpty()) {
-            throw new RuntimeException("Court surface creation failed"); // TODO
+            throw new DatabaseException("Court surface creation failed");
         }
         return courtSurfaceMapper.entityToResponseDto(result.get());
     }
@@ -47,18 +49,25 @@ public class CourtSurfaceServiceImpl implements CourSurfaceService {
         var entityToSave = courtSurfaceMapper.dtoToEntity(surface);
         entityToSave.setId(id);
 
+        if (!surfaceExists(id)) {
+            throw new EntityNotFoundException(String.format("Court surface with id %s not found", id));
+        }
+
         var result = surfaceRepository.updateCourtSurface(id, entityToSave);
         if (result.isEmpty()) {
-            throw new RuntimeException("Court surface update failed"); // TODO
+            throw new DatabaseException("Court surface update failed");
         }
         return courtSurfaceMapper.entityToResponseDto(result.get());
     }
 
     @Override
     public CourtSurfaceDTOs.CourtSurfaceResponseDTO deleteSurface(long id) {
+        if (!surfaceExists(id)) {
+            throw new EntityNotFoundException(String.format("Court surface with id %s not found", id));
+        }
         var result = surfaceRepository.deleteCourtSurface(id);
         if (result.isEmpty()) {
-            throw new RuntimeException("Court surface delete failed"); // TODO
+            throw new DatabaseException("Court surface delete failed");
         }
         return courtSurfaceMapper.entityToResponseDto(result.get());
     }
@@ -67,5 +76,9 @@ public class CourtSurfaceServiceImpl implements CourSurfaceService {
     public List<CourtSurfaceDTOs.CourtSurfaceResponseDTO> getAllSurfaces() {
         var res = surfaceRepository.listCourtSurfaces();
         return res.stream().map(courtSurfaceMapper::entityToResponseDto).toList();
+    }
+
+    private boolean surfaceExists(long id) {
+        return surfaceRepository.getCourtSurfaceById(id).isPresent();
     }
 }
